blender/requirements.txt
numpy==1.13.3
Pillow==4.3.0
scipy==0.19.1
